move customizer's settings and others
