+++ src/Admin/Settings/class-Main.php
@@ -3,7 +3,7 @@
 namespace WP_Plugin_Name\Admin\Settings;
 
 use WP_Plugin_Name\Plugin_Data as Plugin_Data;
-use WP_Plugin_Name\Common\Settings as Common_Settings;
+use WP_Plugin_Name\Common\Settings\Main as Common_Settings;
 use WP_Screen;
 
 // Abort if this file is called directly.
@@ -93,6 +93,7 @@ if ( ! class_exists( Main::class ) ) {
 					'wp-i18n',
 					'wp-api',
 					'wp-components',
+					'wp-data',
 					'wp-element',
 				],
 				Plugin_Data::plugin_version(),
@@ -154,27 +155,8 @@ if ( ! class_exists( Main::class ) ) {
 				wp_die( esc_html__( 'You do not have sufficient permissions to access this page.', Plugin_Data::plugin_text_domain() ) );
 			}
 
-			$link_to_customizer_panel = $this->settings->get_link_to_customizer_panel();
-
 			?>
-			<div class="wrap" id="settings-page">
-				<?php printf( '<h1>%s</h1>', Plugin_Data::get_plugin_display_name() . ' ' . $this->settings->get_settings_word() ); ?>
-
-				<p><?php esc_html_e( "This plugin uses the WordPress Customizer to set its options.", Plugin_Data::plugin_text_domain() ); ?></p>
-				<p><?php esc_html_e( "Click the button below to be taken directly to this plugin's section within the WordPress Customizer.", Plugin_Data::plugin_text_domain() ); ?></p>
-				<p>
-					<?php esc_html_e( "TODO: Add more text here", Plugin_Data::plugin_text_domain() ); ?>
-				</p>
-				<p>
-					<a href="<?php echo esc_url( $link_to_customizer_panel ); ?>"
-					   class="button-primary">
-						<?php esc_html_e( 'Edit Plugin Settings in WP Customizer', Plugin_Data::plugin_text_domain() ) ?>
-					</a>
-				</p>
-				<br><br>
-				<?php
-				?>
-			</div>
+			<div class="wrap" id="settings-page"></div>
 			<?php
 		}
